% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_db.R
\name{get_reference_structure}
\alias{get_reference_structure}
\title{Gets the structure of the reference database (The One That Can Store Them All)}
\usage{
get_reference_structure(outputfile = "webstruct.xlsx",
  reference_source = NULL)
}
\arguments{
\item{outputfile}{Where the structure shall be saved (in Excel format)}

\item{reference_source}{Path to the reference source. Defaults
to the reference shipped with the library.}
}
\value{
Nothing
}
\description{
Gets the structure of the reference database (The One That Can Store Them All)
}
