% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_db.R
\name{get_web_2_xls_dict}
\alias{get_web_2_xls_dict}
\title{Gets the dictionary that translates web database (in its raw form) into the
reference database.}
\usage{
get_web_2_xls_dict(filename = NULL)
}
\arguments{
\item{filename}{Path to the dictionary (if different than the included in the library)}
}
\value{
Returns the dictionary in the form of the data.frame
}
\description{
Gets the dictionary that translates web database (in its raw form) into the
reference database.
}
