% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ONWebDUALSimport.R
\name{importXLSDatabases}
\alias{importXLSDatabases}
\title{Reads in all the Excel databases}
\usage{
importXLSDatabases(filename = NULL, path_prefix = NULL,
  flag_ALSFRS_as_integers = TRUE)
}
\arguments{
\item{filename}{If not-empty, function will save the database into the excel
file with the given name}

\item{path_prefix}{Path containing all the input files. For security reasons the files
are not provided with the library.}
}
\value{
Returns data.table version of the database.
}
\description{
Reads in all the Excel databases
}
