% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ONWebDUALSimport.R
\docType{package}
\name{ONWebDUALSimport}
\alias{ONWebDUALSimport}
\alias{ONWebDUALSimport-package}
\title{Imports patients' records from the web and excel source. Part of the ONWebDUALS project.}
\description{
Imports patients' records from the web and excel source. Part of the ONWebDUALS project.
}
