% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web_db_postprocessing.R
\name{post_processing_for_webdb}
\alias{post_processing_for_webdb}
\title{Various fixes to the web databse after it has been downloaded and put into the reference}
\usage{
post_processing_for_webdb(dbout)
}
\description{
Various fixes to the web databse after it has been downloaded and put into the reference
}
