% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web_documentation_dic.R
\name{read_dictionaries_xlsx}
\alias{read_dictionaries_xlsx}
\title{Reads the web documentation from the external file.}
\usage{
read_dictionaries_xlsx(file_path = "app_dictionaries.xlsx")
}
\value{
Excel documentation in the form of the data.frame
}
\description{
The documentation of the web service has very machine unfriendly format.
It as an Excel table, but the rows represent either entries, or heading.
And sometimes the text inside the cells has more than one line, in which case
there are more than 1 entry for each Excel row.
}
