% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xls_entry_format.R
\name{read_file}
\alias{read_file}
\title{Reads XLS entry format of the database}
\usage{
read_file(filename)
}
\arguments{
\item{filename}{Filename of the database}
}
\value{
List with 2 elements: dataframe containing the read in cases with all the metadata
        and boolean indicating whether the read in file was control or ALS group
}
\description{
Reads XLS entry format of the database
}
