% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_db.R
\name{read_ref}
\alias{read_ref}
\title{Returns empty reference database that is ready to accept new rows}
\usage{
read_ref(row_length = 0, reference_source = NULL,
  flag_ALSFRS_as_integers = TRUE)
}
\arguments{
\item{row_length}{The database may come pre-filled with the
\code{row_length} empty rows. All data will be \code{NA}}

\item{reference_source}{Path to the reference source. Defaults
to the reference shipped with the library.}
}
\description{
Returns empty reference database that is ready to accept new rows
}
