% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ONWebDUALSimport.R
\name{savedb}
\alias{savedb}
\title{Simple function that saves the database in the Excel format}
\usage{
savedb(db, filename)
}
\description{
Simple function that saves the database in the Excel format
}
